+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **Breaking**: Restructuring `Transition` type into `Transition<State, Event>` and `Transition.Result<State>`
 - **Breaking**: Make `Formula` abstract class instead of an interface.
 - **Breaking**: Introducing `Snapshot` type and changing `evaluate` signature.
+- Renaming `Stream` to `Action` (non-breaking change for now)
 
 ## [0.7.0] - June 30, 2021
 - **Breaking**: Remove `events(observable) { }` utility function.